changed dockerization to docker-compose, added main script (TODO make first pretrain script work)
