+++ .dockerignore
@@ -1,0 +1,3 @@
+models
+logs
+.vscode
