--- pretrain-compose.yml
@@ -1,16 +1,0 @@
-
-version: "2"
-
-services:
-
-  chessai-pretrain:
-    build:
-      context: .
-    image: chessai-train
-    command: |
-      python3 chessai/pre_train/pretrain_feature_ext.py &&
-      python3 chessai/pre_train/pretrain_rating.py &&
-      python3 chessai/pre_train/pretrain_model_learning.py
-    volumes:
-      - './models:/app/models'
-      - './logs:/app/logs'
