+++ src/chessai/pretrain/README.md
